add pyproject.toml support
